rename package to autonote
