+++ CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## v0.1.0 (03/01/2023)
 
-- First release of `autodoc`!
+- First release of `autonote`!
