+++ src/autonote/__init__.py
@@ -1,4 +1,4 @@
 # read version from installed package
 from importlib.metadata import version
 
-__version__ = version("autodoc")
+__version__ = version("autonote")
