+++ src/autonote/templates/monthly_report.html.tpl
